pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "pandas>=2.2.3",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "ollama>=0.4.8",
    "requests>=2.32.3",
    "python-dotenv>=1.1.0",
    "ultralytics>=8.3.142",
    "supabase>=2.15.1",
    "pymongo>=4.13.0",
]
